feat: alltests are passed

+++ ewm-service/src/main/java/ru/practicum/ewm/service/service/admin/CategoryAdminServiceImpl.java
@@ -4,21 +4,27 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
-import ru.practicum.ewm.service.dto.category.CategoryDto;
-import ru.practicum.ewm.service.dto.category.NewCategoryDto;
-import ru.practicum.ewm.service.entity.Category;
 import ru.practicum.ewm.service.controller.advice.exception.NotFoundException;
-import ru.practicum.ewm.service.util.mapper.CategoryMapper;
+import ru.practicum.ewm.service.controller.advice.exception.ValidationException;
+import ru.practicum.ewm.service.dto.category.CategoryRequestDto;
+import ru.practicum.ewm.service.dto.category.CategoryResponseDto;
+import ru.practicum.ewm.service.entity.Category;
 import ru.practicum.ewm.service.repository.CategoryRepository;
+import ru.practicum.ewm.service.repository.EventRepository;
+import ru.practicum.ewm.service.util.mapper.CategoryMapper;
 
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class CategoryAdminServiceImpl implements CategoryAdminService {
     CategoryRepository categoryRepository;
+    EventRepository eventRepository;
 
     @Override
-    public CategoryDto addCategory(NewCategoryDto dto) {
+    public CategoryResponseDto addCategory(CategoryRequestDto dto) {
+        if (categoryRepository.existsByName(dto.getName())) {
+            throw new ValidationException("Category name already exist");
+        }
         return CategoryMapper.toDto(
                 categoryRepository.save(
                         CategoryMapper.toEntity(dto)));
@@ -29,14 +35,18 @@ public class CategoryAdminServiceImpl implements CategoryAdminService {
         if (!categoryRepository.existsById(catId)) {
             throw new NotFoundException("no such category");
         }
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ValidationException("category has pinned events");
+        }
         categoryRepository.deleteById(catId);
     }
 
     @Override
-    public CategoryDto patchCategory(Long catId, NewCategoryDto dto) {
+    public CategoryResponseDto patchCategory(Long catId, CategoryRequestDto dto) {
         Category category = categoryRepository.findById(catId)
                 .orElseThrow(() -> new NotFoundException("no such category"));
         category.setName(dto.getName());
-        return CategoryMapper.toDto(category);
+        Category savedCategory = categoryRepository.save(category);
+        return CategoryMapper.toDto(savedCategory);
     }
 }
